attached_assets/requirements_1754111477728.txt
Flask==2.3.2
weasyprint==59.0
google-generativeai==0.4.0
